Add a metrics svc to be able to scrape with prometheus-operator
The additional label app.kubernetes.io/component can separate the exporter
pod from the statefulset ones when used in the svc. There was no change for
the existing labels in order to make it smooth for upgrades. That's why in
the exporter deploy/spec/selector the new label is not added.