Fix kong helm backward compat CRD issue (#2842)
* Fix kong helm backward compat CRD issue

Helm3 deprecated the `crd-install` hook. CRDs instead go in the `crds/` top level directory.

The kong chart must work on both helm2 and helm3 and thus keeps the old template with the `crd-install` hook.
When Helm3 encounters that template it skips it and warns:

```
unsupported hook "crd-install" ignoring
```

We have other 3 charts (e.g. prometheus-operator) that exploit this trick to make the helm chart
compatible with both Helm2 and Helm3.

Whether CRDs are applied or not is controlled by a `--skip-crds` in Helm3 and via a custom value
`ingressController.installCRDs` in the Kong chart (other charts use other names).

However, the `kong` chart defaults `installCRDs` to false. This eliminates the warning
but effectively breaks the chart for Helm2 users who have to figure (possibly by reading the README)
to pass `ingressController.installCRDs=true` during install.

This behavior breaks our backward compatibility guarantees and this PR is a fix for that.

NOTE: our test automation didn't catch that because we only test Helm3 on GKE and we don't treat Helm 2 tests
as release blocking.

* Update Chart.yaml

Co-authored-by: Javier J. Salmerón-García <jsalmeron@bitnami.com>