[bitnami/redis] Redis slave volume init fix (#5059)
* Fix for slave volumePermissions init container command

The command of the volumePermissions init container which is run by the slave pod should by run with `/bin/bash -cx`.

Previously the slave init container execution always failed. After this fix it started working. Tested on minikube.

* Pump PATCH version

* Pump PATCH version of chart version not appVersion

* Dont print evaluated shell commands in initContainer

This makes it more consistent with the other executions of `/bin/sh`

* Consistently use -ec options for shell command in volumePermissions initContainer

* Pump PATCH version of chart version (again after merging master)

* Used /bin/bash -ec consistently for volumePermissions initContainers

Co-authored-by: David Leonhartsberger <david.leonhartsberger@eventworld.com>