[bitnami/kibana] Allow savedObjects to be properly loaded when rewriteBasePath is specified. (#6460)
* kibana savedObjects not processed when server.rewriteBasePath

Address issue #6455 by including the basePath value in the API URL which is used to push saved objects into kibana.

Additionally, add support for ingress.hostname: "*" so that an existing ingress controller can be used to make kibana available.   The current support for ingress.hostname requires a known hostname to be matched by the ingress service.  If the ingress service only has an IP address, then there is now way for the kibnana ingress to be accessed.   The hostname "*" is not a valid specification, so using it to suppress the generation of the hostname field on the ingress, solves the problem without introducing any incompatibilities with the previous default hostname kibana.local.

* Make ingress.yaml hostname optional

Drop hostname if special "*' value is specified.

* Bump kibana Chart version

* Pull in latest bitnami/kibana container with fix for rewriteServerBase

Adopt 7.12.1-debian-10-r24 which has the container side fix for rewriteServerBase.

* Revert "Pull in latest bitnami/kibana container with fix for rewriteServerBase"

This reverts commit d160f01f8e2ef4a45df1b7c101f9addb55b6ff6b.

* [bitnami/kibana] Update components versions

Signed-off-by: Bitnami Containers <containers@bitnami.com>

Co-authored-by: Bitnami Containers <containers@bitnami.com>