[bitnami/argo-cd] Fix SSH repo bug: No user exists for uid 1001 (#9986)
* [bitnami/argo-cd] Fix SSH repo bug: No user exists for uid 1001

Changes the `runAsUser` parameter to 65534 for the repoServer, fixing a problem with SSH based repos not being able to add applications and throwing a `No user exists for uid 1001` error. This is caused due to user 1001 not existing on the repoServer pod. By changing the userID to a user that does exist on the pod, the issue no longer occurs.

Also updates the containerSecurityContext on the controller, repo, server and dex deploys to be in line with Argo CD's own defaults as can be seen [here](https://raw.githubusercontent.com/argoproj/argo-cd/stable/manifests/install.yaml).

Signed-off-by: Kevin Reeuwijk <kevin.reeuwijk@spectrocloud.com>

* [bitnami/argo-cd] Update chart version to 3.1.16

Signed-off-by: Kevin Reeuwijk <kevin.reeuwijk@spectrocloud.com>

* [bitnami/argo-cd] Update metadata

Signed-off-by: Kevin Reeuwijk <kevin.reeuwijk@spectrocloud.com>

* [bitnami/argo-cd] Update metadata

Signed-off-by: Kevin Reeuwijk <kevin.reeuwijk@spectrocloud.com>

* Update README.md with readme-generator-for-helm

Signed-off-by: Bitnami Containers <containers@bitnami.com>

Co-authored-by: Bitnami Containers <containers@bitnami.com>