[bitnami/apache] Fix probes not reading port from .Values (#10286)
* [bitnami/apache] Fix probes not reading port from .Values

From what I can see the  pull request [#1696][1] tried to introduce
both customizable path and port number for the livellyness and readyness
probes, however, the part where it actually read the custom value from
the .Values file was removed before the merge.

With this pull request we actually allow people to change the port, as
else the now default value is ignored and might cause confusion when
nothing happens even though the user has defined a custom port.

Signed-off-by: Jonas Alfredsson <jonasal@axis.com>

[1]: https://github.com/bitnami/charts/pull/1696

* [bitnami/apache] Update components versions

Signed-off-by: Bitnami Containers <containers@bitnami.com>

Co-authored-by: Carlos Rodríguez Hernández <carlosrh@vmware.com>
Co-authored-by: Bitnami Containers <containers@bitnami.com>