[bitnami/rabbitmq] Updated values.yaml (#15934)
* Updated values.yaml

Updated the values.yaml file to prevent loading definition file prior to the cluster being fully initialised. This allows the quorum queues defined within the definition file to setup correctly with followers instead of master with 0 followers. 

Resolves: https://github.com/bitnami/charts/issues/15299

Signed-off-by: Millz12 <millz12@hotmail.com>

* [bitnami/rabbitmq] Updated values.yaml

Updated the values.yaml file to prevent loading definition file prior to the cluster being fully initialised. This allows the quorum queues defined within the definition file to setup correctly with followers instead of only a master with 0 followers.

Signed-off-by: Millz12 <millz12@hotmail.com>

* Update bitnami/rabbitmq/Chart.yaml

Co-authored-by: Ibone González Mauraza <gibone@vmware.com>
Signed-off-by: Millz12 <millz12@hotmail.com>

---------

Signed-off-by: Millz12 <millz12@hotmail.com>
Co-authored-by: Ibone González Mauraza <gibone@vmware.com>