[bitnami/keycloak] switches keycloak-metrics service to point to new port (#30095)
* adds metrics port to values and uses in statefulset, metrics service and network policy

Starting with Keycloak 25 the metrics port was switched from 8080/8443 to 9000. This commit introduces a metrics port in the containerPorts configuration. This metrics port is used to allow ingress via the network policy, expose the container port and also as target for the metrics service. The port of the metrics service itself is not changed. Therefore, metrics are still exposed on port 8080 to the outside. If Keycloak is started with KC_LEGACY_OBSERVABILITY_INTERFACE set, the containerPorts.metrics value can be set to 8080 restoring the old behaviour completely.

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* exposes metrics port only if needed

Only exposes the metrics port if metrics are enabled and the containerPorts.metrics differs from containerPorts.http. In the conditional the ports are quoted as otherwise helm threw an error when setting the containerPorts.metrics value via '-set' about not being able to compare different types.

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* adds documentation for containerPorts.metrics and increments version

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* Update README.md with readme-generator-for-helm

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* removes coalesce

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* updates to next minor version

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* switches metrics service to port 9000

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* casts the port values not to a string but an int

When helm is called with '--set' it the number is interpreted as float whereas the value file interprets it as int.

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* Update README.md with readme-generator-for-helm

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* removes trailing spaces

Signed-off-by: maschmi <maschmi@maschmi.net>

---------

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>
Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>
Signed-off-by: Fran Mulero <francisco-jose.mulero@broadcom.com>
Signed-off-by: maschmi <maschmi@maschmi.net>
Co-authored-by: Bitnami Containers <bitnami-bot@vmware.com>
Co-authored-by: Fran Mulero <francisco-jose.mulero@broadcom.com>