[bitnami/redis] Handle SIGTERM in `kubectl-shared` container (#32085)
* When Redis Sentinel has the master service option enabled, the `kubectl-shared`
container runs a script with an infinite `while+sleep` loop.

This script does not handle SIGTERM and causes the pod to hang upon termination
until the termination grace period elapses.

This change introduces a `preStop` hack to create a `terminate` file, which is used
as a singal to the script that it should exit the infinite loop.

Signed-off-by: Kevin Pullin <kevin.pullin@gmail.com>

* Bump version due to merge conflict

Signed-off-by: Kevin Pullin <kevin.pullin@gmail.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Bot <bitnami.bot@broadcom.com>

---------

Signed-off-by: Kevin Pullin <kevin.pullin@gmail.com>
Signed-off-by: Bitnami Bot <bitnami.bot@broadcom.com>
Co-authored-by: Bitnami Bot <bitnami.bot@broadcom.com>